+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.64.2] - 2021-05-28
+### Fixed
+- Moved aggregate out of full table query for label list [#5584](https://github.com/raster-foundry/raster-foundry/pull/5584)
+
+### Changed
+- Session complete now deactivates previous labels [#5583](https://github.com/raster-foundry/raster-foundry/pull/5583)
+
 ## [1.64.1] - 2021-05-18
 ### Fixed
 - Ensured task actions are updated in single request for labels and task status [#5581](https://github.com/raster-foundry/raster-foundry/pull/5581)
@@ -941,7 +948,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed issue with clamping imagery whose range was greater than, but included values between 0 and 255 [#4177](https://github.com/raster-foundry/raster-foundry/pull/4177)
 - Included missing `pow` operation for decoding json representations of analyses [#4179](https://github.com/raster-foundry/raster-foundry/pull/4140), [#4155](https://github.com/raster-foundry/raster-foundry/issues/4155)
 
-[Unreleased]: https://github.com/raster-foundry/raster-foundry/compare/v1.64.1...HEAD
+[Unreleased]: https://github.com/raster-foundry/raster-foundry/compare/v1.64.2...HEAD
+[1.64.2]: https://github.com/raster-foundry/raster-foundry/compare/v1.64.1...v1.64.2
 [1.64.1]: https://github.com/raster-foundry/raster-foundry/compare/v1.64.0...v1.64.1
 [1.64.0]: https://github.com/raster-foundry/raster-foundry/compare/v1.63.0...v1.64.0
 [1.63.0]: https://github.com/raster-foundry/raster-foundry/compare/v1.62.1...v1.63.0